+++ Cnct/Cnct.Core/Configuration/CloneGitRepositoryTaskSpecification.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
+using System.IO.Abstractions;
 using System.Threading.Tasks;
 using Cnct.Core.Tasks;
 using Newtonsoft.Json;
@@ -8,23 +8,31 @@ using Newtonsoft.Json;
 namespace Cnct.Core.Configuration
 {
     [CnctActionType("cloneGitRepository")]
-    public sealed partial class CloneGitRepositoryTaskSpecification : ICnctActionSpec
+    public sealed partial class CloneGitRepositoryTaskSpecification : CnctActionSpecBase
     {
+        private readonly IFileSystem fileSystem;
         private readonly IGitRunner gitRunner;
 
         public CloneGitRepositoryTaskSpecification()
         {
+            this.fileSystem = new FileSystem();
         }
 
         public CloneGitRepositoryTaskSpecification(IGitRunner gitRunner)
+            : this(gitRunner, new FileSystem())
+        {
+        }
+
+        public CloneGitRepositoryTaskSpecification(IGitRunner gitRunner, IFileSystem fileSystem)
         {
             this.gitRunner = gitRunner;
+            this.fileSystem = fileSystem;
         }
 
         [JsonProperty("repos")]
         public IReadOnlyDictionary<string, string> Repos { get; set; }
 
-        public void Validate()
+        public override void Validate()
         {
             if (this.Repos == null || this.Repos.Count == 0)
             {
@@ -45,21 +53,26 @@ namespace Cnct.Core.Configuration
             }
         }
 
-        public Task ExecuteAsync(ILogger logger, string configDirectoryRoot)
+        public override Task ExecuteAsync(ILogger logger, string configDirectoryRoot)
         {
             var normalizedRepos = new Dictionary<string, string>();
             foreach (var kvp in this.Repos)
             {
                 string dest = kvp.Value.NormalizePath();
-                if (!Path.IsPathRooted(dest))
+                if (!this.fileSystem.Path.IsPathRooted(dest))
                 {
-                    dest = Path.Combine(configDirectoryRoot, dest);
+                    dest = this.fileSystem.Path.Combine(configDirectoryRoot, dest);
                 }
 
                 normalizedRepos[kvp.Key] = dest;
             }
 
-            var cloneTask = new CloneGitRepositoryTask(logger, normalizedRepos, this.gitRunner ?? new ProcessGitRunner(logger));
+            var cloneTask = new CloneGitRepositoryTask(
+                logger,
+                normalizedRepos,
+                this.gitRunner ?? new ProcessGitRunner(logger),
+                this.fileSystem);
+
             return cloneTask.ExecuteAsync();
         }
     }